+++ packages/amplify-frontend-javascript/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.22.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@2.22.3...amplify-frontend-javascript@2.22.4) (2021-06-24)
+
+**Note:** Version bump only for package amplify-frontend-javascript
+
+
+
+
+
 ## [2.22.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-javascript@2.22.2...amplify-frontend-javascript@2.22.3) (2021-06-15)
 
 **Note:** Version bump only for package amplify-frontend-javascript