+++ packages/amplify-java-function-runtime-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@1.8.3...amplify-java-function-runtime-provider@1.8.4) (2021-06-24)
+
+**Note:** Version bump only for package amplify-java-function-runtime-provider
+
+
+
+
+
 ## [1.8.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-java-function-runtime-provider@1.8.2...amplify-java-function-runtime-provider@1.8.3) (2021-06-15)
 
 **Note:** Version bump only for package amplify-java-function-runtime-provider